finish spells UI refactor
