+++ SolastaCommunityExpansion/Viewers/Displays/SpellsDisplay.cs
@@ -1,9 +1,8 @@
-﻿#if false
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using ModKit;
 using SolastaCommunityExpansion.Models;
-using SolastaModApi.Infrastructure;
+using static SolastaCommunityExpansion.Viewers.Displays.Shared;
 
 namespace SolastaCommunityExpansion.Viewers.Displays
 {
@@ -11,56 +10,19 @@ namespace SolastaCommunityExpansion.Viewers.Displays
     {
         private const int SHOW_ALL = -1;
 
-        private const int MAX_COLUMNS = 4;
+        private static HashSet<SpellDefinition> FilteredSpells { get; set; } = SpellsContext.Spells.ToHashSet();
 
-        private const float PIXELS_PER_COLUMN = 225;
+        private static HashSet<SpellDefinition> FilteredSpellsNoCantrips { get; set; } = FilteredSpells.Where(x => x.SpellLevel > 0).ToHashSet();
 
         private static int SpellLevelFilter { get; set; } = SHOW_ALL;
-
-        private static bool Initialized { get; set; }
-
-        private static readonly List<SpellDefinition> SortedRegisteredSpells = new();
-
-        private static readonly List<bool> IsFromOtherModList = new();
-
-        private static bool ExpandAllToggle { get; set; }
-
-        private static readonly Dictionary<string, bool> SpellNamesToggle = new();
-
-        private static string WarningMessage => Main.Settings.AllowDisplayAllUnofficialContent ? ". Spells in " + "brown".color(RGBA.brown) + " were not created by this mod" : string.Empty;
-
-        private static void RecacheSortedRegisteredSpells()
-        {
-            SortedRegisteredSpells.SetRange(SpellsContext.RegisteredSpells
-                .Select(x => x.Key)
-                .Where(x => SpellLevelFilter == SHOW_ALL || x.SpellLevel == SpellLevelFilter)
-                .OrderBy(x => $"{x.SpellLevel} - {x.FormatTitle()}"));
-
-            IsFromOtherModList.SetRange(SpellsContext.RegisteredSpells
-                .Where(x => SpellLevelFilter == SHOW_ALL || x.Key.SpellLevel == SpellLevelFilter)
-                .OrderBy(x => $"{x.Key.SpellLevel} - {x.Key.FormatTitle()}")
-                .Select(x => x.Value.IsFromOtherMod));
-        }
 
         internal static void DisplaySpells()
         {
             int intValue;
             bool toggle;
 
-            if (!Initialized)
-            {
-                SpellsContext.RegisteredSpells.Keys
-                    .Select(x => x.Name)
-                    .ToList()
-                    .ForEach(x => SpellNamesToggle.Add(x, false));
-
-                RecacheSortedRegisteredSpells();
-
-                Initialized = true;
-            }
-
             UI.Label("");
-            UI.Label($". You can individually assign each spell to any caster spell list or simply select the suggested set{WarningMessage}");
+            UI.Label($". You can individually assign each spell to any spell list or simply select the minimum or suggested sets");
 
             if (!Main.Settings.EnableLevel20)
             {
@@ -71,183 +33,98 @@ namespace SolastaCommunityExpansion.Viewers.Displays
 
             using (UI.HorizontalScope())
             {
-                UI.Space(20);
-
-                toggle = SpellsContext.AreAllSpellListsSelected();
+                toggle = SpellsContext.IsAllSetSelected();
 
                 if (UI.Toggle("Select All", ref toggle, UI.Width(PIXELS_PER_COLUMN)))
                 {
-                    SpellsContext.SwitchAllSpellLists(toggle);
+                    SpellsContext.SelectAllSet(toggle);
                 }
 
-                toggle = SpellsContext.AreMinimumSpellListsSelected();
+                toggle = SpellsContext.IsMinimumSetSelected();
                 if (UI.Toggle("Select Minimum", ref toggle, UI.Width(PIXELS_PER_COLUMN)))
                 {
-                    SpellsContext.SwitchMinimumSpellLists(toggle);
+                    if (toggle)
+                    {
+                        SpellsContext.SelectMinimumSet();
+                    }
+                    else
+                    {
+                        SpellsContext.SelectAllSet(false);
+                    }
                 }
 
-                toggle = SpellsContext.AreSuggestedSpellListsSelected();
+                toggle = SpellsContext.IsSuggestedSetSelected();
                 if (UI.Toggle("Select Suggested", ref toggle, UI.Width(PIXELS_PER_COLUMN)))
                 {
-                    SpellsContext.SwitchSuggestedSpellLists(toggle);
-                }
-            }
-
-            UI.Label("");
-
-            using (UI.HorizontalScope())
-            {
-                UI.Space(20);
-
-                intValue = SpellLevelFilter;
-                if (UI.Slider("Spell level filter ".white() + "[-1 to display all spells]".italic().yellow(), ref intValue, SHOW_ALL, 9, SHOW_ALL, "L", UI.AutoWidth()))
-                {
-                    SpellLevelFilter = intValue;
-                    RecacheSortedRegisteredSpells();
-                }
-            }
-
-            UI.Label("");
-
-            if (SortedRegisteredSpells.Count > 0)
-            {
-                using (UI.HorizontalScope())
-                {
-                    UI.Space(20);
-
-                    ExpandAllToggle = SpellNamesToggle.Count == SpellNamesToggle.Count(x => x.Value);
-                    toggle = ExpandAllToggle;
-                    if (UI.Toggle("Expand All", ref toggle, UI.Width(PIXELS_PER_COLUMN)))
+                    if (toggle)
                     {
-                        ExpandAllToggle = toggle;
-                        SpellNamesToggle.Keys.ToList().ForEach(x => SpellNamesToggle[x] = toggle);
+                        SpellsContext.SelectSuggestedSet();
+                    }
+                    else
+                    {
+                        SpellsContext.SelectAllSet(false);
                     }
                 }
             }
-            else
+
+            intValue = SpellLevelFilter;
+            if (UI.Slider("spell level filter ".bold().italic().white() + "[-1 to display all spells]".bold().italic().yellow(), ref intValue, SHOW_ALL, 9, SHOW_ALL))
             {
-                UI.Label($". No unofficial level {SpellLevelFilter} spells available".red().bold());
-            }
+                SpellLevelFilter = intValue;
 
-            UI.Label("");
-
-            for (var i = 0; i < SortedRegisteredSpells.Count; i++)
-            {
-                var spellDefinition = SortedRegisteredSpells[i];
-                var spellName = spellDefinition.Name;
-                var spellTitle = $"{spellDefinition.SpellLevel} - {spellDefinition.FormatTitle()}";
-
-                if (IsFromOtherModList[i])
+                if (intValue == SHOW_ALL)
                 {
-                    spellTitle = spellTitle.color(RGBA.brown);
+                    FilteredSpells = SpellsContext.Spells.ToHashSet();
+                    FilteredSpellsNoCantrips = FilteredSpells.Where(x => x.SpellLevel > 0).ToHashSet();
                 }
-
-                toggle = SpellNamesToggle[spellName];
-                if (UI.DisclosureToggle(spellTitle.yellow(), ref toggle, 200))
+                else
                 {
-                    SpellNamesToggle[spellName] = toggle;
-                }
-
-                if (SpellNamesToggle[spellName])
-                {
-                    DisplaySpell(spellDefinition);
-                }
-
-                UI.Label("");
-            }
-        }
-
-        internal static void DisplaySpell(SpellDefinition spellDefinition)
-        {
-            UI.Label("");
-            UI.Label($". {spellDefinition.FormatDescription()}");
-
-            using (UI.HorizontalScope())
-            {
-                UI.Space(20);
-
-                using (UI.VerticalScope())
-                {
-                    DisplaySpellListSelection(spellDefinition);
-                }
-            }
-        }
-
-        internal static void DisplaySpellListSelection(SpellDefinition spellDefinition)
-        {
-            var spellName = spellDefinition.Name;
-            bool toggle;
-            int columns;
-            var current = 0;
-            var spellListsTitles = SpellsContext.SpellLists.Keys;
-            var spellLists = SpellsContext.SpellLists.Values;
-            var spellListsCount = spellLists.Count;
-
-            UI.Label("");
-
-            using (UI.HorizontalScope())
-            {
-                toggle = SpellsContext.AreAllSpellListsSelected(spellDefinition);
-                if (UI.Toggle("Select All", ref toggle, UI.Width(PIXELS_PER_COLUMN)))
-                {
-                    SpellsContext.SwitchAllSpellLists(toggle, spellDefinition);
-                }
-
-                toggle = SpellsContext.AreMinimumSpellListsSelected(spellDefinition);
-                if (UI.Toggle("Select Minimum", ref toggle, UI.Width(PIXELS_PER_COLUMN)))
-                {
-                    SpellsContext.SwitchMinimumSpellLists(toggle, spellDefinition);
-                }
-
-                toggle = SpellsContext.AreSuggestedSpellListsSelected(spellDefinition);
-                if (UI.Toggle("Select Suggested", ref toggle, UI.Width(PIXELS_PER_COLUMN)))
-                {
-                    SpellsContext.SwitchSuggestedSpellLists(toggle, spellDefinition);
+                    FilteredSpells = SpellsContext.Spells.Where(x => x.SpellLevel == SpellLevelFilter).ToHashSet();
+                    FilteredSpellsNoCantrips = FilteredSpells.Where(x => x.SpellLevel > 0).ToHashSet();
                 }
             }
 
             UI.Label("");
 
-            while (current < spellListsCount)
-            {
-                columns = MAX_COLUMNS;
+            UI.Div();
 
-                using (UI.HorizontalScope())
+            foreach (var kvp in SpellsContext.SpellLists)
+            {
+                var spellListDefinition = kvp.Value;
+                var name = spellListDefinition.name;
+                var displayToggle = Main.Settings.DisplaySpellListsToggle[name];
+                var sliderPos = Main.Settings.SpellListSliderPosition[name];
+                var spellEnabled = Main.Settings.SpellListSpellEnabled[name];
+
+                void AdditionalRendering()
                 {
-                    while (current < spellListsCount && columns > 0)
+                    toggle = SpellsContext.SpellListContextTab[spellListDefinition].IsMinimumSetSelected;
+                    if (UI.Toggle("Select Minimum", ref toggle, UI.Width(PIXELS_PER_COLUMN)))
                     {
-                        var spellListDefinition = spellLists.ElementAt(current);
-                        var spellListName = spellListDefinition.Name;
-                        var spellListTitle = spellListsTitles.ElementAt(current);
+                        SpellsContext.SpellListContextTab[spellListDefinition].SelectMinimumSet();
+                    }
 
-                        if (spellDefinition.SpellLevel > spellListDefinition.MaxSpellLevel
-                            || (spellDefinition.SpellLevel == 0 && !spellListDefinition.HasCantrips))
-                        {
-                            current++;
-                            continue;
-                        }
-
-                        toggle = Main.Settings.SpellSpellListEnabled[spellName].Contains(spellListName);
-                        if (UI.Toggle(spellListTitle, ref toggle, UI.Width(PIXELS_PER_COLUMN)))
-                        {
-                            if (toggle)
-                            {
-                                Main.Settings.SpellSpellListEnabled[spellName].Add(spellListName);
-                            }
-                            else
-                            {
-                                Main.Settings.SpellSpellListEnabled[spellName].Remove(spellListName);
-                            }
-
-                            SpellsContext.SwitchSpellList(spellDefinition, spellListDefinition);
-                        }
-
-                        columns--;
-                        current++;
+                    toggle = SpellsContext.SpellListContextTab[spellListDefinition].IsSuggestedSetSelected;
+                    if (UI.Toggle("Select Suggested", ref toggle, UI.Width(PIXELS_PER_COLUMN)))
+                    {
+                        SpellsContext.SpellListContextTab[spellListDefinition].SelectSuggestedSet();
                     }
                 }
+
+                DisplayDefinitions(
+                    $"{kvp.Key}:".yellow(),
+                    SpellsContext.SpellListContextTab[spellListDefinition].Switch,
+                    spellListDefinition.HasCantrips ? FilteredSpells : FilteredSpellsNoCantrips,
+                    spellEnabled,
+                    ref displayToggle,
+                    ref sliderPos,
+                    AdditionalRendering);
+
+                Main.Settings.DisplaySpellListsToggle[name] = displayToggle;
+                Main.Settings.SpellListSliderPosition[name] = sliderPos;
             }
+
+            UI.Label("");
         }
     }
 }
-#endif
